backend/requirements.txt
# Dependências principais da API
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Banco de dados
mysql-connector-python==8.2.0
pymysql==1.1.0

# Processamento de PDFs e documentos
pdfplumber==0.10.3
pandas==2.1.4
PyMuPDF==1.23.8
tabula-py==2.8.2
google-generativeai==0.3.2
Pillow==10.1.0
openpyxl==3.1.2
pytesseract==0.3.10
pdf2image==1.16.3

# Utilitários
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.5.0

# Logging e monitoramento
loguru==0.7.2

# CORS e segurança
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4